third_party/WebKit/LayoutTests/platform/mac/fast/dom/Element/scrollTop-scrollLeft-body-expected.txt

Test for scrollTop/Left values of scrollable document.body in quirks and strict modes

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS iframeStrict1.contentDocument.documentElement.scrollTop is 0
PASS iframeStrict1.contentDocument.documentElement.scrollLeft is 100
FAIL iframeStrict1.contentDocument.body.scrollTop should be 200. Was 187.
PASS iframeStrict1.contentDocument.body.scrollLeft is 0
PASS iframeStrict1.contentWindow.pageYOffset is 0
PASS iframeStrict1.contentWindow.pageXOffset is 100
PASS iframeQuirks1.contentDocument.documentElement.scrollTop is 0
PASS iframeQuirks1.contentDocument.documentElement.scrollLeft is 0
FAIL iframeQuirks1.contentDocument.body.scrollTop should be 200. Was 187.
PASS iframeQuirks1.contentDocument.body.scrollLeft is 0
PASS iframeQuirks1.contentWindow.pageYOffset is 0
PASS iframeQuirks1.contentWindow.pageXOffset is 0
PASS iframeStrict2.contentDocument.documentElement.scrollTop is 0
PASS iframeStrict2.contentDocument.documentElement.scrollLeft is 100
PASS iframeStrict2.contentDocument.body.scrollTop is 0
PASS iframeStrict2.contentDocument.body.scrollLeft is 0
PASS iframeStrict2.contentWindow.pageYOffset is 0
PASS iframeStrict2.contentWindow.pageXOffset is 100
PASS iframeQuirks2.contentDocument.documentElement.scrollTop is 0
PASS iframeQuirks2.contentDocument.documentElement.scrollLeft is 0
PASS iframeQuirks2.contentDocument.body.scrollTop is 0
PASS iframeQuirks2.contentDocument.body.scrollLeft is 0
PASS iframeQuirks2.contentWindow.pageYOffset is 0
PASS iframeQuirks2.contentWindow.pageXOffset is 0
PASS iframeStrict3.contentDocument.documentElement.scrollTop is 0
PASS iframeStrict3.contentDocument.documentElement.scrollLeft is 100
FAIL iframeStrict3.contentDocument.body.scrollTop should be 200. Was 172.
PASS iframeStrict3.contentDocument.body.scrollLeft is 0
PASS iframeStrict3.contentWindow.pageYOffset is 0
PASS iframeStrict3.contentWindow.pageXOffset is 100
PASS iframeQuirks3.contentDocument.documentElement.scrollTop is 0
PASS iframeQuirks3.contentDocument.documentElement.scrollLeft is 0
FAIL iframeQuirks3.contentDocument.body.scrollTop should be 200. Was 172.
PASS iframeQuirks3.contentDocument.body.scrollLeft is 0
PASS iframeQuirks3.contentWindow.pageYOffset is 0
PASS iframeQuirks3.contentWindow.pageXOffset is 0
PASS iframeStrict4.contentDocument.body.scrollTop is 0
PASS iframeStrict4.contentDocument.body.scrollLeft is 0
PASS iframeQuirks4.contentDocument.body.scrollTop is 0
PASS iframeQuirks4.contentDocument.body.scrollLeft is 0
PASS successfullyParsed is true

TEST COMPLETE
           
